Fixed the compiler errors intrudoced by changing the RelationalOperators class, and fixed bugs where the IRConstFloat expected integer instead of double and where the logical expression used the wrong types for the IRJump constructor

+++ src/ir/expressions/ir_const_float.cc
@@ -1,5 +1,5 @@
 #include "ir_const_float.hh"
 
-IRConstFloat::IRConstFloat(int in) : value(in)
+IRConstFloat::IRConstFloat(double in) : value(in)
 {
 }
